Attempt to fix a bug: Added no cache headers
- Added headers to nginx configuration files to try and prevent browsers caching sites. There's a possible bug where if a site is using a framework that sets a domain name environment variable in a .env file such as the `WP_HOME` for Laravel Bedrock, then when trying to request the site and the TLD is different from the one set in Valet, then the site automatically redirects to use the domain from the environment variable, which ends in Valet returning a 404 error, because as far as Valet is concerned it's not valid site.

Example: `WP_HOME='http://mySite.test'`, Valet gets a request to `http://mySite.dev`, the site will auto redirect to `http://mySite.test`.

The real issue is when this 404 response is cached by the browser and keeps requesting the cached version of the site even if the TLD has been changed to match. The no cache headers attempt to prevent this. Though testing has been inconclusive to whether this makes a difference or not.

- Added some documentation for the `share` command.
- Various auto-formatting.
- Added a known issue for this "bug".