Update Silly dependency and change `start` command
Previously, Silly had a bug where the `runCommand` wouldn't pass any arguments to the sub-command that it was going to run. As a work around, Valet had to just use `passthru` to explicitely run the sub-command in the commandline. With the new Silly update, the `runCommand` is fixed so we can just run the sub-command (with arguments) internally instead.