Fix: The underlying cause to #3
After lots of testing, the underlying cause to #3 (which was partially fixed in #4) is due to a rogue `PhpCgiXdebug::installed()` in the `install` command function, which was inadvertently caused by a copy/paste situation and forgot to change it the `$xdebug` variable that checks if `--xdebug` option is set. The erroneous function call was null and void since Xdebug hasn't been installed (yet).

- Changed function call to variable and some dev comments.