Fix: Xdebug `restart` function
If there is a version of Xdebug installed and other versions of PHP didn't have their corresponding Xdebug installed, and the `restart` function of `PhpCgiXdebug` class is ran, then Valet would spit out an error in complaint of the Xdebug PHP version failing to restart, because it's not installed.

This is fixed by adding an if statement in the core `restart` function of the `PhpCgi` class (of which `PhpCgiXdebug` class shares) to check if the Xdebug service is installed. The function would only restart the service if it is installed.