Fix: Bug with nginx trying to stop twice and encountering a rare WMI error.
During usage of the `valet use` command, Valet explicitly stops the nginx service with `winsw->stop()`, and then uninstalls and reinstalls the service (to make nginx reconfigure to the new default PHP version). The problem is that during uninstalling, the `uninstall` function of the `WinSW` class (of which Nginx `uninstall` calls) also runs the `stop()` function to stop the service, even though the service is already stopped.

Then on a very rare occasion, a WMI fatal error is outputted:

FATAL - WMI Operation failure: InvalidServiceControl
WMI.WmiException: InvalidServiceControl
   at WMI.WmiRoot.BaseHandler.CheckError(ManagementBaseObject result)
   at WMI.WmiRoot.InstanceHandler.Invoke(Object proxy, MethodInfo method, Object[] arguments)
   at WinSW.Program.<Run>g__Stop|2_3(<>c__DisplayClass2_0& )
   at WinSW.Program.Run(String[] argsArray, IWinSWConfiguration descriptor)
   at WinSW.Program.Main(String[] args)

Though scripts thereafter seem to continue running without problems. Not 100% sure if this error is related to the service trying to stop twice, as it only occurs on 1% of command runs. The other 99% it outputs a nice message "The service with ID 'valet_nginx' is not running".

However, to try and combat the useless WMI error, the `uninstall` function of WinSW now only stops the service if it is running, and will also work for other services that needs to be uninstalled.

- Added `isRunning` function to check if the service is running, and use it in the `uninstall` function to determine if `stop` should be called.
- Removed redundant argument from the `stop` call because the function doesn't take any arguments.