Fix: The server port replacing bug.
Further tests reveal that the deprecated commented out code in commit a0dc616 is not needed, and removing it fixes a bug. But the bug described in the commit is the wrong bug.

The bug that this actually fixes is where a non-secured isolated site's server port was being replaced with the PHP port and was being served with the default server instead of the isolated server.

So instead of the usual port 80: `127.0.0.1:80` in the site's isolated config file, nginx was actually listening for the site on port 9002: `127.0.0.1:9002`, which is the port where PHP 7.4 was configured to be hosted. This confused nginx because it couldn't find port 80 in the site's config file, so it used the default server instead, in which the default PHP version was set as 8.1 on the port 9001, and thus serving the site under the wrong PHP version.

- Removed the `preg_replace` in the `replacePhpVersionInSiteConf` function of the `Site` class.
(This doesn't affect the PHP isolation and still works as intended.)
- Updated Changelog.