Fix: `diagnose` command copy diagnostics not actually copying.
For some unknown reason, the copying of the diagnostics to the clipboard just stopped working. To fix this, changed the cli type to powershell to ensure that the copy command `clip` is available.

Also changed where the diagnostics text file are saved to for copying. They now save to the valet home directory `~/.config/valet`. Otherwise, it'll try to save to the current working directory of the terminal, which could potentially be a protected directory like `Program Files`.