feat: add emergency batch script
Errors will occurs when running `composer global update` without uninstalling valet first, because the services are still installed and therefore the files cannot be removed and updated since they're in-use.  The batch script is to emergency stop and uninstall the services, and so a subsequent composer update is able to work.

- Added the .bat file to stop and uninstall services.
- Updated README to document this.
- Added a copy command in `Configuration::install` method to copy the .bat file into the /.config/valet directory for safe keeping.