fix: `Ngrok::hasAuthToken` method to actually check if there's an auth token property set when the config file exists.
-   Fixed `hasAuthToken` to explicitly check for an `authtoken` property when the config file exists.

Currently, the method only checks if the config file exists, if it doesn't then the authtoken isn't set. But what if the user added the config manually or edited the file. The method would then always `return true` even though the token might not exist.

Fixed by adding an explicit check of the contents if the file does exist. The contents of the yml config file is read and converted to an associative array.
From this we can then check if the config version is 2 or 3 (v3 has a new config format), and then we can check if the `authtoken` key exists. In v3 the `authtoken` is defined in a separate `agent` array, so the checks need to be adjusted.

- Added `symfony/yaml` dependency to convert contents of a yml to an associative array.

-   Refactored `getNgrokConfig` method to optionally return only the path of the config file (without the leading `--config` cli flag). Required for the fixed `hasAuthToken` method.