Revert "refactor: Ngrok::start method to test ngrok for errors before actually running the command for real."
This reverts commit f8be62b264f15e8e2df9be8b0970c6de6b9d511c.

The fix doesn't work properly, because if ngrok doesn't have any errors, the command will be executed as the actual command. After a while Symfony's Process may decide to give up on the request and spit out a "timeout" error. There doesn't seem to be a a good way to make this work, so reverting is the best option at the moment.