refactor: `Ngrok::start` method to output a message if ngrok errors that the executable is "too old"
Because we can't test ngrok commands for errors or errors in the executable, we need to return the output of any errors and check the output for a specific error code relating to the "too old" error (ERR_NGROK_121). Then output a message (along with the original error output) to the user to inform them that they can update ngrok executable and also upgrade the config file themselves by performing the valet commands.

All errors will still be output to terminal, but we need to make sure if this specific error occurs, the user knows what to do.

- Removed the functionality of the share command starting a new CMD window, since a new window will surpress any errors. We need to be able to output the errors as and when they happen.

- Added various output messages to the terminal, so it's more descriptive as to what's happening.

- Added new `CommandLine::shellExec` method to execute the command in the terminal, and also be able to return the output as a string to the calling method. This is just a wrapper around the native PHP `shell_exec` function.