fix: `Filesystem::symlinkAsUser` method to create a real symbolic link.
As per the docs [1], `mklink` params are `/D` for symbolic link, `/H` for hard link, and `/J` for directory junction.

Currently, site links are created as a directory junction. This is not a real symbolic link, and PHP's `is_link` function always fails. To fix, we use the proper param `/D` to create a real symlink.

[1] https://learn.microsoft.com/en-us/windows-server/administration/windows-commands/mklink