feat: add new `Upgrader::convertJunctionsToSymlinks` method
- Add new `Filesystem::convertJunctionsToSymlinks` method to remove the old Windows junction links, and re-link the sites as real symlinks.

- Add new `Filesystem::getJunctionLinks` method to get all the site links that are junctions.

This uses the Windows `dir` command to get details about all the linked directories, and piping the output to `findstr` command to narrow down the selection to only those that are defined as a `JUNCTION`. We then obtain the link name and the full site path from the details to be used in `convertJunctionsToSymlinks` method.

- Add `convertJunctionsToSymlinks` method in the `Upgrader` to call the `Filesystem::convertJunctionsToSymlinks` only if a config key is false/not present.

The config key prevents the upgrade running again, since it's only a one-time upgrade.