fix: a weird bug when unzipping a zip file would have a strange path with a duplicate directory.
During unzipping the downloaded Ansicon zip file, an error would be thrown because of a weird path with a duplicated `ansicon` directory (`\ansicon\ansicon`).

The error is:

   valet-windows\bin\ansicon\ansicon"' is not recognized as an internal or external command,
operable program or batch file.

When the path is ran behind the scenes of Windows `tar` command, the path is tried to run as command.

Not sure why the `ansicon` directory is duplicated or why the path is trying to be run a command while unzipping.

However, using the `run` method instead of `passthru` somehow fixes the issue.