fix: `Filesystem::move` method errors when moving files to a directories that don't exist.
As per a comment (https://www.php.net/manual/en/function.copy.php#62807) on the PHP `copy` function docs, `copy` doesn't create directories if they don't exist in the path. It only copies to a fully resolved path with existing directories.

So if the extracted directories have multi-level subdirectories, then an error will occur:
"PHP Warning:  copy(path): Failed to open stream: No such file or directory"

Fixed by removing both the class `copy` and `unlink` methods, and replacing with the PHP `rename` function (https://www.php.net/manual/en/function.rename.php). This function renames a filepath and moves it to a new location. The filepath can be a file or directory. If it's is a directory, the contents will be moved at the same time. If the destination directory or any subdirectories does not exist, it will be created.
The `rename` function is the PHP docs recommended way for moving files: "If you wish to move a file, use the rename() function." (https://www.php.net/manual/en/function.copy.php)