fix: `Site::isolate` and `Site::unisolate` methods spamming the `Nginx::stop` and `Nginx::restart` methods
The `Nginx::stop` and `Nginx::restart` methods are spammed if the `Site::isolate` or `Site::unisolate` methods are used in a loop, causing a plethora of Windows UAC popups (if `sudo` wasn't used).

Fixed by removing the `Nginx` calls within the `Site` methods and added the `Nginx::restart` method call to the relevant command definitions in valet.php after where `Site::isolate` or `Site::unisolate` are called.