fix: phpcgi installation can't `put` the php service file into the stubs directory.
In commit d101a63, all references the directory `stubs` was replaced with the `Filesystem::getStub` method. This included the path to the stubs directory when `put`ting a file to the directory. This broke the `PhpCgi::installService` method because `getStubs` is getting the contents of the existing file, not getting the stub filepath.

Fixed by removing the `getStub` call and adding back the stubs directory path.