fix: `Diagnose::editOutput` method param type hint error `Expected type 'object'. Found 'string'`
A subtle error that only occurs in vscode editor (and not when running valet); it is caused by setting the `$output` param's type hint to `string` in the docblock. But in actual fact this could be either `string` or an instance of the `ProcessOutput` class.

Fixed by adding `ProcessOutput` to type hint it in the docblock param.