fix: `PHP Fatal error:  Uncaught Error: Class "Specific/{driver}" not found` error.
When using the deprecated legacy drivers, a PHP fatal error occurs for all legacy drivers, saying that the `Specific` respective driver class that the legacy driver extends can not be found. Extending the namespaced class is supposed to prevent this using the PSR-4 autoloading, but autoloading doesn't seem to work here.

To fix, we need to `require` all of the `Specific` namespaced drivers before the legacy drivers are `require`d. This then ensures all namespaced classes are available in the legacy drivers, and fixes the PHP class not found error.