feat: add `packageExeName` property to allow a different exe name to the package name.
- Added new `packageExeName` protected property to specify the name of it's executable file.

- Changed `packageExe` method to use the new `packageExeName` property if it is defined otherwise use the `packageName`.

This is needed because if an executable has a different name to that of the package name, we need to be able to tell the class that.

Certain exe files can't/shouldn't be renamed just in case changing the name affects its functionality, if it's own name is referenced in the exe's code.