feat: remove `Acrylic` executables in favour of downloading latest from SourceForge.
Removed:

- Removed all `Acrylic` bin files.

Added:

- Added new `Acrylic` class with the namespace `Valet\Packages`, and extending the base `Package` class. This class will never clash with the normal `Acrylic` class because of the different namespaces.

Because SourceForge doesn't have a proper API for releases, we have to use the `best_release.json` endpoint to obtain the latest `.exe` version. But because the json file only lists the download URL of the standalone `.exe` file, we have to replace the filename to get the download URL for the portable zip version (`Acrylic-Portable.zip`).

- Added new `Acrylic_IPv4_Binding_Address.ini` stub file to add the custom IPv4 binding localhost loopback address as fixed in PR https://github.com/yCodeTech/valet-windows/pull/15.

- Added a `changeLocalIpv4BindingAddress` method to replace the default value of `LocalIPv4BindingAddress`  with our custom value from the new stub file.