feat: enable nginx to display http status 500 error page
If a PHP error occurs internally while trying to access a site, nginx fails silently, and the browser just displays it's generic 500 error page. It is rather difficult to determine what caused the error from this.

To help users debug the error, we can tell nginx to redirect the user to a custom 500 error page, which tells them where to look for the error.

Added:

-  Added new `500.html` template as the error page which describes the log file to look for the nginx error, and also use the valet command.

Changed:

- Enabled nginx to intercept php fastcgi errors in all site conf stubs, and redirect to the new 500.html page.

- Changed the replacement code for the site conf stubs to now replace any `VALET_ERROR_TEMPLATE_PATH` constants with the real absolute path of the error template location.